# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SwiftEntry API",
        "description": "Backend REST de SwiftEntry \u2014 plataforma de venta de boletos para eventos.\n\n**Autenticaci\u00f3n:** todas las rutas protegidas requieren un JWT en el header `Authorization: Bearer <token>`. Obt\u00e9n el token en `POST /swift_entry/auth/login` y p\u00e9galo en el bot\u00f3n **Authorize** de arriba.\n\n**Flujo principal:** Evento \u2192 Localidad \u2192 Asiento \u2192 Reserva \u2192 Pago \u2192 Boleto\n",
        "contact": {
            "name": "SwiftEntry Team",
            "email": "gerardo.port1107.psql@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/swift_entry/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar todos los roles",
                "description": "Solo ADMINISTRATOR",
                "operationId": "getAllRoles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear rol",
                "operationId": "creatRole",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/swift_entry/reservations": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Listar todas las reservas",
                "description": "Solo ADMINISTRATOR",
                "operationId": "getAllReservations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Crear reserva",
                "description": "Aparta los asientos seleccionados por 15 minutos. M\u00e1ximo 5 asientos por reserva",
                "operationId": "createReservation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener reserva por ID",
                "description": "Solo ADMINISTRATOR",
                "operationId": "getReservationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Cancelar reserva",
                "description": "Libera todos los asientos de la reserva",
                "operationId": "cancelReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/reservations/{reservationId}/seats/{reservationSeatId}": {
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Quitar un asiento de la reserva",
                "description": "Si era el \u00faltimo asiento, cancela la reserva completa",
                "operationId": "removeSeatFromReservation",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "reservationSeatId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {}
                }
            },
...
            "RoleRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 0
                    }
                },
                "required": [
                    "description",
                    "name"
                ]
            },
            "ReservationRequestDTO": {
                "type": "object",
                "properties": {
                    "localitySeatIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "maxItems": 5,
                        "minItems": 0
                    }
                },
                "required": [
                    "localitySeatIds"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}